+++ src/controls/mdfbDatePicker.js
@@ -1,7 +1,7 @@
 'use strict'
 
 var skipLogic = require('../utils/skipLogic')()
-var asynchValidator = require('../validators/asynchValidator')()
+var asyncValidator = require('../validators/asyncValidator')()
 
 module.exports = function ($compile) {
   return {
@@ -21,7 +21,7 @@ module.exports = function ($compile) {
       }
 
       if (scope.field.validation) {
-        asynchValidator.init(scope, scope.field, scope.globals)
+        asyncValidator.init(scope, scope.field, scope.globals)
       }
     }
   }
